fix(issues): Remove similar issues padding
Splits similar issues into SimilarIssues and SimilarIssuesTab which has the page layout padding.
